Added compatibility with relationl sluggable unicity field if the relation has i18n + inheritance behaviors
Imagine we have a schema like this:

Foo:
  actAs:
    I18n:
      fields: [name]
      actAs:
        Sluggable:
          [...]
          fields: [name]
          uniqueBy: [%TRANSLATION_OWNER_CLASS%.master_id, lang]
  columns:
    id: [...]
    master_id: [...]

FooBar:
  inheritance:
    extends: Foo
    type: column_aggregation
    [...]

FooBaz:
  inheritance:
    extends: Foo
    type: column_aggregation
    [...]

You want a slug unicity on the i18n record with the lang and the master_id of the main component (Foo). However, with inheritance you can't define all relation names (FooBar.master_id, FooBaz.master_id), so you need to use the special %TRANSLATION_OWNER_CLASS% keyword which replace the right component name for each subclass for you.
This is probably not the most elegant way to achieve this, but I couldn't figure another method in a short time :/