Fixed regeneration of config_autoload.yml.php at each request made by doctrine
Some doctrine behaviors like i18n uses a class "generator" to generate
on the fly the *Translation classes when the table is loaded.

The problem is this generator runs a class_exists() on the *Translation class
before generate it, which triggers the symfony autoloading process. When
the class has not yet been generated at this point (which occurs most of
time), symfony re-walk ALL files and rebuild a fresh
config_autoload.yml.php.

By adding "false" to the 2nd argument of class_exists() we avoid
autoload mechanism to be triggered.